# Conflicts:
#   apps/faq/models/question_answer.py
#   apps/information/models/photo_gallery_models.py
#   apps/information/serializers/news.py
#   apps/information/serializers/photo_gallery_serializers.py
#   apps/information/urls.py
#   apps/information/views/photo_gallery_views.py
#   config/settings/base.py
#   config/urls.py
#   config/wsgi.py

+++ apps/information/models/management.py
@@ -22,7 +22,7 @@ class Management(DateTimeMixin):
         null=True,
         verbose_name='Должность'
     )
-    start_year = models.IntegerField(
+    start_year = models.PositiveSmallIntegerField(
         verbose_name='Год начала работы'
     )
     birth_date = models.DateField(
